tasks have been completed

+++ src/main/java/lang/print/gaps/finalModuleTask/TemperatureConverter.java
@@ -2,7 +2,7 @@ package lang.print.gaps.finalModuleTask;
 
 public class TemperatureConverter {
     public void toFahrenheit(int temperatureCelsius){
-        int temperatureFahrenheit = temperatureCelsius * 9 / 5 + 32;
+        float temperatureFahrenheit = temperatureCelsius * 9 / 5 + 32f;
         System.out.println(temperatureFahrenheit);
     }
 }
